fix(observe): filter version in stats SQL and split SDK fetch helper
Apply version_name in readStatsSB before the row limit, and move fetchObserve out of the CLI command module so SDK consumers do not import Clack UI.
